new module to handle gw_xvx algorithms, cleaner this way
